Add multi-channel support to `updateBrowser.js`. Pull common code into `common.js`
Fixes https://github.com/brave/browser-laptop/issues/11795

Auditors: @aekeus
